Product Context - Fixes #5
