+++ src/components/ProductList.jsx
@@ -1,13 +1,19 @@
 import ProductCard from "./ProductCard";
+import { useProducts } from "../context/ProductContext";
 
-const ProductList = ({products}) => {
+const ProductList = () => {
+    const {products, loading, error} = useProducts();
+
     return ( 
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 gap-6">
+          {loading && <p>Loading...</p>}
+          {error && <p className="text-red-500">{error}</p>}
           {products.map((product) => (
             <ProductCard key={product.id} product={product} />
           ))}
         </div>
+
      );
 }
- 
+
 export default ProductList;
