+++ src/main.jsx
@@ -2,9 +2,13 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.jsx'
+import { ProductProvider } from './context/ProductContext.jsx'
+
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <App />
+    <ProductProvider>
+      <App />
+    </ProductProvider>
   </StrictMode>,
 )
